fixed login errors
