--- src/main/java/com/example/project/models/gameStates/CurrentGameState.java
@@ -1,8 +1,0 @@
-package com.example.project.models.gameStates;
-
-public class CurrentGameState
-{
-    private LevelModel level;
-
-
-}
